+++ src/DataDictionary/DataDictionary.css
@@ -10,12 +10,14 @@
 
 .data-dictionary__left {
   width: var(--data-dictionary__left-width);
+  min-width: var(--data-dictionary__left-width);
   height: 100%;
   background-color: var(--g3-color__white);
 }
 
 .data-dictionary__right {
   width: calc(100vw - var(--data-dictionary__left-width));
+  min-width: 900px;
   overflow-y: scroll;
 }
 
@@ -49,6 +51,7 @@
   width: 50%;
   text-align: center;
   padding: 10px;
+  outline: none;
 }
 
 .data-dictionary__switch-button:hover {
