+++ src/DataDictionary/DataModelStructure/DataModelStructure.css
@@ -1,5 +1,5 @@
 .data-model-structure {
-  border-top: 1px solid var(--g3-color__lightgray);
+  border-top: 1px solid var(--g3-color__silver);
 }
 
 .data-model-structure__header {
