+++ src/DataDictionary/graph/Canvas/Canvas.css
@@ -11,6 +11,7 @@
   margin-bottom: 10px;
   text-align: center;
   cursor: pointer;
+  outline: none;
 }
 
 .canvas__zoom-button:hover,
