+++ src/DataDictionary/graph/Legend/Legend.css
@@ -60,6 +60,7 @@
   height: 8px;
   background-color: var(--g3-color__gray);
   cursor: pointer;
+  outline: none;
 }
 
 .data-dictionary-graph-legend__close:hover {
@@ -73,6 +74,7 @@
   border-radius: 50%;
   display: block;
   background-color: var(--g3-color__base-blue);
+  outline: none;
 }
 
 .data-dictionary-graph-legend__info-icon {
