+++ src/DataDictionary/graph/NodePopup/NodePopup.css
@@ -62,6 +62,7 @@
   height: 8px;
   background-color: var(--g3-color__gray);
   cursor: pointer;
+  outline: none;
 }
 
 .node-popup__close:hover {
