+++ src/DataDictionary/graph/OverlayPropertyTable/index.js
@@ -5,7 +5,8 @@ import OverlayPropertyTable from './OverlayPropertyTable';
 const ReduxOverlayPropertyTable = (() => {
   const mapStateToProps = state => ({
     hidden: state.ddgraph.overlayPropertyHidden,
-    node: state.ddgraph.highlightingNode ? state.submission.dictionary[state.ddgraph.highlightingNode.id] : null,
+    node: state.ddgraph.highlightingNode
+      ? state.submission.dictionary[state.ddgraph.highlightingNode.id] : null,
   });
 
   const mapDispatchToProps = dispatch => ({
