+++ src/components/layout/NavBar.less
@@ -23,7 +23,7 @@
 }
 
 .nav-bar__header--full-width {
-  width: inherit;
+  min-width: 100vw;
   padding: 0 20px;
 }
