+++ src/components/layout/TopBar.less
@@ -12,7 +12,7 @@
 }
 
 .top-bar__header--full-width {
-  width: inherit;
+  min-width: 100vw;
   padding: 0 20px;
 }
