+++ src/index.less
@@ -1,12 +1,12 @@
 .protected-content {
   width: 1220px;
   margin: auto;
-  min-height: ~"calc(100vh - 221px)"; /* 40px topbar, 81px navbar, 100px footer */ 
+  min-height: ~"calc(100vh - 221px)";
   background-color: var(--g3-color__bg-cloud);
 }
 
 .protected-content--dd-page {
   width: 100%;
+  min-width: 1220px;
   height: calc(~"100vh - 121px");
-  overflow: hidden;
 }
